# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AttestationOut": {
        "additionalProperties": false,
        "description": "A signed lineage bundle, for export. S28, AC-F11.",
        "properties": {
          "artefact": {
            "description": "What this attests to.",
            "title": "Artefact",
            "type": "string"
          },
          "complete": {
            "description": "Whether the chain reaches licensed roots with no gaps.",
            "title": "Complete",
            "type": "boolean"
          },
          "gaps": {
            "description": "Every gap, where there are any.",
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "issuedAt": {
            "description": "When the bundle was produced.",
            "format": "date-time",
            "title": "Issuedat",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "description": "The canonical bundle that was signed.",
            "title": "Payload",
            "type": "object"
          },
          "payloadSha256": {
            "description": "Digest of the canonical bundle.",
            "title": "Payloadsha256",
            "type": "string"
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The signature, when the chain is complete. An incomplete chain is exported unsigned and says so: signing an attestation over a gap would certify the gap.",
            "title": "Signature"
          },
          "siteId": {
            "description": "Which site issued it.",
            "title": "Siteid",
            "type": "string"
          }
        },
        "required": [
          "artefact",
          "complete",
          "issuedAt",
          "siteId",
          "payload",
          "payloadSha256"
        ],
        "title": "AttestationOut",
        "type": "object"
      },
...
      "LedgerEntryDetailOut": {
        "additionalProperties": false,
        "description": "One ledger entry with its payload and its chain links. S27.",
        "properties": {
          "actor": {
            "description": "Who caused it.",
            "title": "Actor",
            "type": "string"
          },
          "entryHash": {
            "description": "This entry's hash.",
            "title": "Entryhash",
            "type": "string"
          },
          "id": {
            "description": "Entry identifier.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "description": "The signed payload.",
            "title": "Payload",
            "type": "object"
          },
          "prevHash": {
            "description": "The previous entry's hash.",
            "title": "Prevhash",
            "type": "string"
          },
          "recomputedHash": {
            "description": "The hash recomputed here from prev_hash and payload.",
            "title": "Recomputedhash",
            "type": "string"
          },
          "seq": {
            "description": "Position in this site's chain.",
            "title": "Seq",
            "type": "integer"
          },
          "siteId": {
            "description": "Whose segment this belongs to.",
            "title": "Siteid",
            "type": "string"
          },
          "subjectId": {
            "description": "Which one.",
            "title": "Subjectid",
            "type": "string"
          },
          "subjectType": {
            "description": "What kind of thing it concerns.",
            "title": "Subjecttype",
            "type": "string"
          },
          "transition": {
            "description": "The state transition recorded.",
            "title": "Transition",
            "type": "string"
          },
          "ts": {
            "description": "When it was recorded.",
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "verified": {
            "description": "Whether the recomputed hash equals the recorded one. Computed on read rather than trusted: an entry viewer that displays a stored hash proves nothing.",
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "siteId",
          "seq",
          "prevHash",
          "entryHash",
          "ts",
          "actor",
          "subjectType",
          "subjectId",
          "transition",
          "payload",
          "recomputedHash",
          "verified"
        ],
        "title": "LedgerEntryDetailOut",
        "type": "object"
      },
...
      "LedgerSlice": {
        "additionalProperties": false,
        "description": "A slice of the ledger, with its chain verification.",
        "properties": {
          "divergence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where the chain first diverges, if it does.",
            "title": "Divergence"
          },
          "items": {
            "description": "The entries in the slice.",
            "items": {
              "$ref": "#/components/schemas/LedgerEntryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Page size.",
            "title": "Limit",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cursor for the next page.",
            "title": "Nextcursor"
          },
          "verified": {
            "description": "Whether the returned slice's hash chain verifies end to end.",
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "limit",
          "verified"
        ],
        "title": "LedgerSlice",
        "type": "object"
      },
      "LineageOut": {
        "additionalProperties": false,
        "description": "A lineage attestation. AC-F11.",
        "properties": {
          "approval": {
            "additionalProperties": true,
            "description": "The approval, carrying the sole approver exception (SAD 9.4).",
            "title": "Approval",
            "type": "object"
          },
          "artefact": {
            "description": "The artefact this chain describes.",
            "title": "Artefact",
            "type": "string"
          },
          "complete": {
            "description": "Whether the chain reaches licensed roots with no gaps.",
            "title": "Complete",
            "type": "boolean"
          },
          "corpusHashes": {
            "description": "Every source and corpus hash in the chain.",
            "items": {
              "type": "string"
            },
            "title": "Corpushashes",
            "type": "array"
          },
          "gaps": {
            "description": "Every gap, where there are any.",
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "licences": {
            "description": "Every licence in the chain.",
            "items": {
              "type": "string"
            },
            "title": "Licences",
            "type": "array"
          },
          "nodes": {
            "description": "The chain itself.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "artefact",
          "complete"
        ],
        "title": "LineageOut",
        "type": "object"
      },
...
      "Problem": {
        "description": "An RFC 9457 problem document.",
        "properties": {
          "code": {
            "description": "Machine readable, stable problem code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation for this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URI of this occurrence.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nThe DRAUPNIR control plane API.\n\nErrors are RFC 9457 problem documents served as `application/problem+json`.\nMutating endpoints require an `Idempotency-Key`. Collections are cursor\npaginated. Mutable resources carry an `ETag` and require `If-Match`.\nLong operations return 202 with a run identifier; nothing blocks on training.\n",
    "summary": "CIM-56 model factory control plane",
    "title": "DRAUPNIR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ledger": {
      "get": {
        "description": "A slice of this site's ledger segment, verified end to end.\n\n    The verification travels with the slice. An auditor who has to recompute it\n    themselves will not, and an endpoint that returned entries without saying\n    whether they chain is an endpoint that makes tampering look like data.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "getLedger",
        "parameters": [
          {
            "description": "Opaque cursor from a previous page's `nextCursor`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's `nextCursor`.",
              "title": "Cursor"
            }
          },
          {
            "description": "First sequence to return.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First sequence to return.",
              "title": "From"
            }
          },
          {
            "description": "Last sequence to return.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last sequence to return.",
              "title": "To"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 200,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size.",
              "title": "Limit"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerSlice"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Ledger slice with chain verification",
        "tags": [
          "audit"
        ]
      }
    },
    "/v1/ledger/{entry_hash}": {
      "get": {
        "description": "The payload, the actor, and both hashes. S27.\n\n    The response carries a hash recomputed here from `prev_hash` and the\n    canonical payload, beside the one that was stored. An entry viewer that\n    rendered only the stored hash would prove nothing: the stored hash is\n    exactly what a tamperer would have rewritten.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "getLedgerEntry",
        "parameters": [
          {
            "description": "The entry's own hash.",
            "in": "path",
            "name": "entry_hash",
            "required": true,
            "schema": {
              "description": "The entry's own hash.",
              "pattern": "^[0-9a-f]{64}$",
              "title": "Entry Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "One ledger entry, with its hash recomputed",
        "tags": [
          "audit"
        ]
      }
    },
    "/v1/lineage/{artefact}": {
      "get": {
        "description": "The complete chain to base model licences and corpus hashes. AC-F11.\n\n    `complete` and `gaps` are returned rather than an error on an incomplete\n    chain: an auditor asking about a broken lineage needs to see where it\n    breaks, and a 404 or a 500 tells them nothing. Producing a signed\n    *attestation* over a broken chain is what is refused, and that refusal is\n    in SKIDBLADNIR.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "getLineage",
        "parameters": [
          {
            "description": "The artefact's SHA-256.",
            "in": "path",
            "name": "artefact",
            "required": true,
            "schema": {
              "description": "The artefact's SHA-256.",
              "pattern": "^[0-9a-f]{64}$",
              "title": "Artefact",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Full lineage attestation",
        "tags": [
          "audit"
        ]
      }
    },
    "/v1/lineage/{artefact}/attestation": {
      "get": {
        "description": "The lineage as a canonical, hashed bundle. S28, AC-F11.\n\n    An incomplete chain exports **unsigned**, and says so. Signing an\n    attestation over a gap would certify the gap, which is worse than refusing\n    to sign: a signature is read as a statement that somebody checked, and\n    nobody checked what is missing.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "exportAttestation",
        "parameters": [
          {
            "description": "The artefact's SHA-256.",
            "in": "path",
            "name": "artefact",
            "required": true,
            "schema": {
              "description": "The artefact's SHA-256.",
              "pattern": "^[0-9a-f]{64}$",
              "title": "Artefact",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttestationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "A signed lineage bundle, for export",
        "tags": [
          "audit"
        ]
      }
    },
...
  }
}